% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalytics_operations.R
\name{kinesisanalytics_delete_application_input_processing_configuration}
\alias{kinesisanalytics_delete_application_input_processing_configuration}
\title{Deletes an InputProcessingConfiguration from an input}
\usage{
kinesisanalytics_delete_application_input_processing_configuration(
  ApplicationName, CurrentApplicationVersionId, InputId)
}
\arguments{
\item{ApplicationName}{[required] The Kinesis Analytics application name.}

\item{CurrentApplicationVersionId}{[required] The version ID of the Kinesis Analytics application.}

\item{InputId}{[required] The ID of the input configuration from which to delete the input
processing configuration. You can get a list of the input IDs for an
application by using the DescribeApplication operation.}
}
\description{
Deletes an InputProcessingConfiguration from an input.
}
\section{Request syntax}{
\preformatted{kinesisanalytics$delete_application_input_processing_configuration(
  ApplicationName = "string",
  CurrentApplicationVersionId = 123,
  InputId = "string"
)
}
}

\keyword{internal}
