% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalytics_operations.R
\name{kinesisanalytics_describe_application}
\alias{kinesisanalytics_describe_application}
\title{Returns information about a specific Amazon Kinesis Analytics
application}
\usage{
kinesisanalytics_describe_application(ApplicationName)
}
\arguments{
\item{ApplicationName}{[required] Name of the application.}
}
\description{
Returns information about a specific Amazon Kinesis Analytics
application.
}
\details{
If you want to retrieve a list of all applications in your account, use
the ListApplications operation.

This operation requires permissions to perform the
\code{kinesisanalytics:DescribeApplication} action. You can use
\code{DescribeApplication} to get the current application versionId, which
you need to call other operations such as \code{Update}.
}
\section{Request syntax}{
\preformatted{kinesisanalytics$describe_application(
  ApplicationName = "string"
)
}
}

\keyword{internal}
