% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalytics_operations.R
\name{kinesisanalytics_stop_application}
\alias{kinesisanalytics_stop_application}
\title{Stops the application from processing input data}
\usage{
kinesisanalytics_stop_application(ApplicationName)
}
\arguments{
\item{ApplicationName}{[required] Name of the running application to stop.}
}
\description{
Stops the application from processing input data. You can stop an
application only if it is in the running state. You can use the
DescribeApplication operation to find the application state. After the
application is stopped, Amazon Kinesis Analytics stops reading data from
the input, the application stops processing data, and there is no output
written to the destination.
}
\details{
This operation requires permissions to perform the
\code{kinesisanalytics:StopApplication} action.
}
\section{Request syntax}{
\preformatted{kinesisanalytics$stop_application(
  ApplicationName = "string"
)
}
}

\keyword{internal}
